requirements.txt
streamlit==1.28.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0
joblib==1.3.2
nltk==3.8.1
matplotlib==3.7.2
PyPDF2==3.0.1
python-docx==0.8.11
seaborn>=0.12.0